[External] [stdlib] Add struct `UnsafeMaybeUninitialized` (#41185)
[External] [stdlib] Add struct `UnsafeMaybeUninitialized`

This struct is private at the moment and for internal use only. Once we
are sure that it's ready for public use, we'll call it
`MaybeUninitialized`. Heavily borrowed from
https://doc.rust-lang.org/stable/std/mem/union.MaybeUninit.html

This can allow us to add a destructor to `InlineArray` without impacting
`InlineList` and SBO. We can make `InlineList` use a
`InlineArray[MaybeUninitialized[T]]`.

@ConnorGray for review since you have a good idea of what to do with it.
See
https://github.com/modularml/mojo/issues/2869#issuecomment-2137962617
for background.

Three things to note:
1) I didn't measure the performance impact of this struct compared to
working directly with uninitialized memory. Even if there is a
performance penalty right now, do we really care at this point since we
don't have benchmarks yet? Even if the implementation could be faster,
this PR gives us unit tests and a struct with method signatures to
iterate on.
2) This struct is private, I didn't want to put too much pressure on
this PR to have something perfect on the first try. Furthermore, we
don't have good docstrings yet, those can be added later on.
3) I use
```
`__mlir_type[`!pop.array<`, _size.value, `, `, Self.ElementType, `>`]`
```
but this may not be the best MLIR type to do this. I'm not very
experienced with MLIR, so if there is something more appropriate that we
can use, let me know.

---------

Co-authored-by: Gabriel de Marmiesse <gabriel.demarmiesse@datadoghq.com>
Co-authored-by: Connor Gray <code@connorgray.com>
Closes modularml/mojo#2888
MODULAR_ORIG_COMMIT_REV_ID: ef173db08ff7b602ab7dc66aa4009acd69c64df8