[External] [stdlib] Use `ValueDestructorRecorder` in `test_inline_list.mojo` (#42108)
[External] [stdlib] Use `ValueDestructorRecorder` in
`test_inline_list.mojo`

This PR is a small piece of mojo#2965 . I'm trying to make mojo#2965 smaller to
make it easier to work on.

In itself, this PR is a refactoring. The goal is to unify the
"destructor counters" that are present in the unit tests and use only
the one in `stdlib/test/test_utils/types.mojo`. The goal of those
structs is to record the number of `__del__` call being made to make
sure the lifecycle of elements in a collection is correct. To make it
usable with `InlineList`, I had to add the explicit copy constructor.

Co-authored-by: Gabriel de Marmiesse <gabriel.demarmiesse@datadoghq.com>
Closes modularml/mojo#3094
MODULAR_ORIG_COMMIT_REV_ID: 4ff2d0ea67c29b144b6c76d3fb0f8db93143e41a