[External] [stdlib] Add the trait `CollectionElementNew` to `List` (#42499)
[External] [stdlib] Add the trait `CollectionElementNew` to `List`

At first I wanted to make a second constructor but it was not possible
because of this bug:
* https://github.com/modularml/mojo/issues/3137

so in the end, I replaced the existing one. Hopefully end users won't
get too many surprises with that, as it can change the type of the
result when doing things like `List(List[Int]())` since it uses then the
constructor with variadic arguments.

Co-authored-by: Gabriel de Marmiesse <gabriel.demarmiesse@datadoghq.com>
Closes modularml/mojo#3138
MODULAR_ORIG_COMMIT_REV_ID: f2e100d916ae7cefc49933583e390502ab732522