[External] [stdlib] Move `find()` from `StringRef` to `StringSlice` (#43931)
We are just moving the StringRef implementation of `find`. We'll delete
`StringRef.find()` later on. The base implementation is here:
https://github.com/modularml/mojo/blob/nightly/stdlib/src/utils/stringref.mojo#L497

`String.find()` is now plugged to `StringSlice.find()`

In general we are trying to move away from `StringRef` and at some point
maybe delete it in favor of `StringSlice` which is safer, both in terms
of memory and in terms of utf-8 validity.

Co-authored-by: Gabriel de Marmiesse <gabriel.demarmiesse@datadoghq.com>
Closes modularml/mojo#3228
MODULAR_ORIG_COMMIT_REV_ID: 973e4efceee95f5a35b47aff1ec8bf3f5a4e9e21