[External] [stdlib] Make `InlineArray` call its elements' destructor (#42451)
[External] [stdlib] Make `InlineArray` call its elements' destructor

Fix this issue:
*  https://github.com/modularml/mojo/issues/2869

A little explanation of what is going on here, this should make the
review easier.

### Modifications to `InlineArray`
`InlineArray` made use of the `@value` decorator, which generates, in
addition to a constructor, the following methods: `__moveinit__`,
`__copyinit__` and `__del__`. While it's easy to understand what is
generated when working with attributes that are values, it's less clear
what is happenning when working with plain MLIR types like we do here.

The next step is to manually define those to make sure they do what we
want. **We will always assume that the `InlineArray` is completely
initialized**. Thus when copying, moving or deleting, we must copy, move
and delete each element in the `InlineArray` too.

To make the implementation easier, instead of using an `UnsafePointer`
and its methods to control the lifecycle of the values
(`move_from_pointee` & friends), we fill the array with
`UnsafeMaybeUninitialized`. This struct has methods that allows us to
control the lifecycle without necessarily work with pointers.

### What do we do with `InlineArray[...](unsafe_uninitialized=True)`?

This method is a massive footgun. This is because throughout the methods
of `InlineArray`, we always assume that the values are initialized. So
doing any moving, copying or deleting with uninitialized memory is UB.

This is not much of an issue for very trivial types where moving and
copying are just copying bits and the destructor does nothing. This can
still cause UB if the user uses some uninitialized memory but it's fine
if the caller is careful. For example, in the stdlib, we use
`unsafe_uninitialized` to store some `UInt8` here and there and it's
fine.

As such, the constructor was kept, but an extensive docstring was added
to explain in which situation it is okay to use it.

### How `InlineArray` is now used by `InlineList`

`InlineList` wants to carefully manage the lifecycle of each element in
the `InlineArray` while still making the `InlineArray` believe that it
owns the lifecycle of its elements. In those kind of situation,
`UnsafeMaybeUninitialized` is the right struct to use.

So we use an `InlineArray[UnsafeMaybeUninitialized[ElementType]]` inside
`InlineList`. I know it can be confusing since `InlineArray` itself also
uses `UnsafeMaybeUninitialized` but this shouldn't be visible from
outside `InlineArray` so it should be fine. Basically it disables all
lifecycle methods, and `__del__` is a no-op.

This allows `InlineList` to manually control the lifecycle of each
elements which has an index below the size (above the size of the
`InlineList`, it's truly uninitialized memory).

### Some guideline for the code review:

I advise to read the files in this order:

* The files `stdlib/test/test_utils/types.mojo` and
`stdlib/test/collections/test_inline_list.mojo` have been refactored to
make the tests of `InlineArray` use the `ValueDestructorRecorder`. This
could have been a separate PR but I was too lazy to split this PR even
further.
* The file `stdlib/test/utils/test_tuple.mojo` has a new test that
ensure that all the elements in the `InlineArray` have their destructors
called.
* The file `stdlib/src/utils/static_tuple.mojo` has all the changes that
makes `InlineArray` use `UnsafeMaybeUninitialized` internally. While we
could have technically use pointers, it makes writing code easier when
controlling lifecycle.
* Finally, we switch the internal type of `InlineList` in
`stdlib/src/collections/inline_list.mojo`. We now use
`InlineArray[UnsafeMaybeUninitialized[ElementType]]` which allows the
`InlineList` to control exactly the lifecycle. I repeat here: yes in can
be confusing because overall, we use two `UnsafeMaybeUninitialized` on
top of each other when using
`InlineArray[UnsafeMaybeUninitialized[ElementType]]`, but the code is
still readable, because one "belongs" to `InlineArray` and one "belongs"
to `InlineList.`

### Modifications from internal review
- The `InlineArray` ability to call element destructors is hidden behind
a parameter and defaulted to `False` for compatibility in IR generation
internally. We'll progressively move these over to use `True` and then
later, hopefully remove the parameter entirely.

Co-authored-by: Gabriel de Marmiesse <gabriel.demarmiesse@datadoghq.com>
Closes modularml/mojo#2965
MODULAR_ORIG_COMMIT_REV_ID: 7332d9a3f1dc5b21d257da5d1e58294537c5d38c