[stdlib] Introduce prelude module
Currently, the `builtin` module (and all of its transitive imports from
every mojo file within the `builtin` directory) is auto-imported into every
Mojo program.
This is bad for a few reasons:
1. It couples the implementation location of an entity with where it
ought to be exported.
2. It's confusing that there are *additional* things auto-imported into
every Mojo program that *do not* live in the `builtin` directory.

This manifests in user confusion at best since other stdlib modules such
as `random` are imported into user's programs without them even knowing.
This is because the imports of implementation files were also transitively
included in this builtin set of auto imports. Specifically, prior to this change,
the transitive closure set of auto-imported modules was:

```
{'memory', 'sys', 'os', 'utils', 'python', 'bit', 'random', 'math', 'builtin', 'collections'}
```

This is not great - that set represents most of the standard library!

Attempt to start clarifying these gotchas by introducing a new module:
`prelude`, which has no definitions itself, but re-exports the fixed,
enumerated set of entities that we want to be automatically imported into every
Mojo program. Currently, this is roughly the set of public (not starting with
an `_`) entities in the `builtin` module today, plus a few load-bearing
common imports. These load-bearing ones will be chipped away at shortly after
and removed soon, but are nice to allow for progressively landing this work
while already fixing all of the missing import issues throughout the codebase.

MODULAR_ORIG_COMMIT_REV_ID: af800eb391dec3ad535e0d1bbbd185e23717bb7e

Signed-off-by: Manuel Saelices <msaelices@gmail.com>