[stdlib] Add proposal for hardware time stamp counter support
Hardware time stamp counters (TSCs) are present on most modern
architectures and offer a cheaper way to get timestamps than other
methods. This is useful for benchmarking, logging, and more exotic
usecases.

This proposal includes an API which should meet most needs, as well as
how to handle the absence of a reliable TSC.