[mojo-lang] Allow implicit declaration of variables in `fn`s
Mojo and Python both allow implicit declaration of variables within
a `def`.  This patch changes the compiler to allow them within an
`fn` as well.

The rationale here is that we haven't seen significant safety or
correctness benefit from requiring explicit variable definitions,
and this change makes the language more consistent and ergonomic.
We do want systems programmers to use `fn` to signify that
functions don't raise, and providing more strict handling of
arguments (for example, `fn` statements still strictly scope bindings
in `with` blocks).  As such, we want `def` and `fn` to be able to
be different, just not in this dimension.

Mojo still supports `var` declarations and we have no plans to remove
them.  They are useful in cases where you want to reuse the same name
for different purposes in different parts of a function, and are
required for
struct member definitions.

MODULAR_ORIG_COMMIT_REV_ID: dc8d115339ea2a792fbd78d0bacda37a536abd74

Signed-off-by: Manuel Saelices <msaelices@gmail.com>