[mojo-lang] Implement exclusivity checking, currently as a warning.
This enhances the Mojo compiler to notice when mutable values are
aliased in
function calls, diagnosing them with a warning (which needs to be
upgraded to
an error).

Mojo (and
[Rust](https://doc.rust-lang.org/beta/book/ch04-02-references-and-borrowing.html)
and [Swift](https://swift.org/blog/swift-5-exclusivity/)) require that
all mutable values have
unique ownership to provide memory safety, and also performance
benefits. This has been
long-planned and this patch finally implements it.

This is being phased in as a warning, because there are a couple of
non-trivial uses of
aliasing pointers that need to be sorted in the standard library,
notably in the
`_stable_sort` and `_quicksort` implementations.

This fixes https://github.com/modularml/mojo/issues/1734 and possibly
others.

MODULAR_ORIG_COMMIT_REV_ID: cd9c6b3c45d4eea98fbc2a743a015c0cd80f2dee

Signed-off-by: Manuel Saelices <msaelices@gmail.com>