[mojo-lang] Revise memory transfer implementation. (#45488)
This rewrites the implementation of the transfer operator (`x^`) for
memory values. Instead of generating a `lit.transfer_mem_ownership`,
just treat the value as an RValue, allowing it to be directly
consumed by whatever it is passed to.  This is better in a number
of ways:

1) Simplicity: this eliminates `lit.transfer_mem_ownership` and
   supporting logic.
2) IR compactness: this removes the op, and a new temp lifetime.
3) Correctness: it isn't correct to generate a new lifetime referring
   to the same memory location!

The correctness issue was shown when exclusivity checking didn't notice
exclusivity violations with the move operation (as the testcase now
shows).  However, it is conceivable that one could construct situations
where both the previous and the new lifetime were reused, which would
have been a miscompilation.

MODULAR_ORIG_COMMIT_REV_ID: 5a0d7aad39f2d9ae6779dfbf1bd1c4d7bf0e285c

Signed-off-by: Manuel Saelices <msaelices@gmail.com>