[stdlib] Reduce code size bloat in `debug_assert`
Since `debug_assert` is not a macro, if the condition has a side-affect, the
code generated persists in the final executable and cannot be DCEd by the
compiler.  So, in practice, we see a lot of calls to `String.format` (which was
called from inside `debug_assert` which does not-so-small amount of work to
construct a `String` that may never used at runtime in a program.

In
[Internal link]
`debug_assert` was loosened to work on `Formattable` to be consistent with
`print` and `abort`, but given that this isn't free in terms of code size bloat,
walk this back a bit to work just on `Stringable` types.  We could loosen it
even further to `StringLiteral`, but currently the stdlib is passing a runtime
`String` to `debug_assert` in a few places, so don't go quite that far yet.

MODULAR_ORIG_COMMIT_REV_ID: 379a607cb982fa035834a43acfea7ba18108cefc