[External] [stdlib] Add new method `SIMD._dynamic_shuffle()` (#45892)
[External] [stdlib] Add new method `SIMD._dynamic_shuffle()`

This method is incredibly useful to implement fast UTF-8 validation.
It's used a lot in the paper [Validating UTF-8 In Less Than One
Instruction Per Byte](https://arxiv.org/abs/2010.03090). It notably
allowed me to implement a fast utf8 validation 6 times faster than the
one currently in the stdlib. Since it's a vectorized dynamic
vector-lookup, it can be useful in many other scenarios.

Note that I also tried to implement it with calling the llvm intrinsic
but I would get those kind of errors:

```mojo
    return sys.llvm_intrinsic[
        "llvm.shufflevector",
        SIMD[DType.uint8, 16],
        has_side_effect=False,
    ](lookup_table, indices)
```

`could not find LLVM intrinsic: "llvm.shufflevector"` or `error: could
not find LLVM intrinsic: "shufflevector"` if I remove the llvm prefix.
If that's possible to use it, guidance would be appreciated.

The current implementation works, even if it's not the fastest for some
vector sizes and arch. So I believe if we can't find anything better
quickly, we can merge it and iterate on it later.

Co-authored-by: Gabriel de Marmiesse <gabrieldemarmiesse@gmail.com>
Closes modularml/mojo#3397
MODULAR_ORIG_COMMIT_REV_ID: e2b3095316825a6994ad1c4e023727fa53cbebb6

Signed-off-by: Manuel Saelices <msaelices@gmail.com>