Initial support for nix shell/develop
Nix shell/develop provides similar functionality to 'magic shell' but works with the
Nix package manager (https://nixos.org/download/) and is generally used to manage
an entire operating system.
Nix does not follow the FHS standard which means that magic does not work on systems
using Nix out of the box and providing such a developer environment makes it easier
for people using Nix to use and contribute to the software provided by Modular.
On a system with a default Nix installation the shell can be entered by going to the
mojo repository path and executing 'nix-shell'.
With flake support enabled it also works with 'nix develop'.

Signed-off-by: Thomas Mader <thomas.mader@gmail.com>