[stdlib] Add alias for void* in ffi
To improve expressiveness/intent, add an alias for C's `void*` when
interacting
in FFI places.  Apply this in the CPython parts of the stdlib.

It can get applied to other parts of the stdlib in a follow-up.

MODULAR_ORIG_COMMIT_REV_ID: f90cd1fb352687ec98f74135bf78ab7bb682597e