[mojo-lang] Put `StringLiteral` into a new global lifetime.
This introduces compiler support for a new `#lit.static.lifetime`
lifetime
to model things that have static storage duration, e.g. string literals
and
global variables.

This adopts it in the standard library in the `StringLiteral` type (and
also
cleans up some other stuff w.r.t. it), replacing the use of
`AnyLifetime`
which could point to anything.

This is better for alias analysis and subtyping in general, but is
important
because a subsequent patch will change the behavior of `AnyLifetime`
and I don't want every use of `print` to be impacted by it :)

MODULAR_ORIG_COMMIT_REV_ID: a1a072ee2e77b827bda304055616bbdbcba4b2a1