[mojo-lang] Introduce an escape hatch for exclusivity checking.
This introduces a new `__unsafe_disable_nested_lifetime_exclusivity`
decorator which disables nested lifetime exclusivity checking.  This
is used on functions that take lifetimes nested in one or more arguments
but guarantees they do not actually access anything in that lifetime.

As its long and scary name indicates, this is not intended to be
generally
used, but is an important escape hatch for now. In the future, we're
likely
to replace this with a safe (checked) mechanism to say that nested
lifetimes are not accessed.  I haven't implemented such logic yet and
don't want to be blocked on it.

This also adds this to the methods in `Reference`, `Span` and
`StringSlice` that can use these, and re-introduces some unit tests
that I removed when switching exclusivity to an error.

MODULAR_ORIG_COMMIT_REV_ID: e8783c56545dea0b3dada5756e0815d3d3b1d2cb