[stdlib] Fix PyGILState_* signatures
`PyGILState_Ensure` returns a `PyGILState_STATE` which is an opaque
"handle" to the thread state, not a bool. Adjust the wrapper calls and
the dual function: `PyGILState_Release` so it accepts the handle rather
than the bool.  Note the `PyGILState_STATE` is actually an enum in C
and is four bytes (an `int`).  So, model it as a struct with a value that is
four bytes for ABI compatibility purposes.

Getting these right will be important as we aim to fix our "respect the
GIL" issues in python interop.

MODULAR_ORIG_COMMIT_REV_ID: 542fcf89b830298314ce47629029861b946fa4d6