[stdlib] Make `merge` private in sort module
`merge` is a bit of a dangerous API since it takes a span as an
arguments and expects its arguments to be uninit data. It's currently
quite specific to its one caller. So, make it private to avoid confusion
that it's more general than it is right now.

While in the neighborhood, remove an unused variable.

MODULAR_ORIG_COMMIT_REV_ID: 1ad4d1ebaea7f41998f25fa181c14522502af71f