[Proposal] Stdlib insider docs
People working on the standard library need to have some more
information about the API contracts and behavior of the runtime and compiler
builtins in order to be able to write correct and performant code in the
stdlib. This is not intended to be high effort documentation or
acknowledgement that a feature is stable, but simply communication of
enough information for those of us without access to the compiler source
and autogenerated mlir documentation to work with them correctly.

Signed-off-by: Owen Hilyard <hilyard.owen@gmail.com>