Convert Mojo docs from notebooks to MDX
This PR is purely a file format change; there are no copy changes.

We started writing the docs as notebooks because it was convenient to
write code snippets that we could execute in situ and print outputs for
the docs. We also thought this would allow us to test all the code in
CI, but that has proven much more difficult than we had hoped. Although
it's still nice to run code in situ, this does not outweigh the
problems. The primary problems are: Editing the docs is slightly more
cumbersome compared to markdown because it's done in a GUI, and GitHub
reviews are very challenging because we can't comment on the rendered
view (you must instead open the JSON source and make comments there,
which is thoroughly confusing).

Switching to markdown should make editing and reviewing documents much
faster. For code testing, we will rely more on GitHub code examples that
_can_ be tested in CI (and copy them into the markdown).

We're using `.mdx` instead of `.md` because MDX is more reliable and
extensible with Docusaurus. The CommonMark processor for `.md` files has
some bugs, such as an inability to faithfully parse some types of HTML
inside markdown.

MODULAR_ORIG_COMMIT_REV_ID: 595d2f763197aa502263eb76ce388b0781738366