[mojo-lang] Fix a crash with weird llvm dialect ops.
This fixes the parser to not assume that the attribute returned by
the MLIR fold hook for an op returns a TypedAttr.  If it doesn't, we
can just retain the op as it is.  The LLVM dialect has an
`llvm.mlir.zero` op that folds to `#llvm.zero`.

This fixes https://github.com/modularml/mojo/issues/3805

MODULAR_ORIG_COMMIT_REV_ID: e2b18182eea2529afd721b3e3969ed3f36d2c3ff