[stdlib] Skip trivial destructors in `List`
If the elements are hinted as "trivial", skip running the destructors
for them. Note that we don't have a checked equivalent of
`std::is_trivially_destructible`, so this is just relying on the hint
provided to `List`. For non-trivial types, we still run the destructors
on each of the elements, of course.

Future PRs will take advantage of the `hint` parameter in `copyinit`.

MODULAR_ORIG_COMMIT_REV_ID: c92f98b69f1a40b7b671b6392a7b71b54cfb61b8